requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
